+++ Assets/Testing/ArrayListTester.cs
@@ -6,14 +6,14 @@ using Gravinium.Jilwer.Core.Collections;
 [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
 public class ArrayListTester : UdonSharpBehaviour
 {
-    public JilwerRuntime jilwer;
+    [HideInInspector] public JilwerRuntime jilwer;
 
-    private ObjectArrayList _list;
+    private ArrayList _list;
     private int _counter = 1;
 
     private void Start()
     {
-        _list = ObjectArrayList.New(jilwer);
+        _list = ArrayList.New(jilwer);
     }
 
     public override void Interact()
@@ -24,7 +24,8 @@ public class ArrayListTester : UdonSharpBehaviour
         string msg = "[";
         for (int i = 0; i < _list.Length(); i++)
         {
-            msg += _list.Get(i) + (i < _list.Length() - 1 ? ", " : "]");
+            if (_list.TryGet(i, out object num) != Error.None) { return; }
+            msg += num + (i < _list.Length() - 1 ? ", " : "]");
         }
         Debug.Log("[ArrayListTester] " + msg);
     }