+++ Packages/org.gravinium.jilwer/Runtime/Scripts/Core/Error.cs
@@ -1,0 +1,9 @@
+﻿namespace Gravinium.Jilwer.Core
+{
+    public enum Error
+    {
+        None,
+        IndexOutOfBounds,
+        InvalidType,
+    }
+}